Added more abstraction
