+++ src/Diagnostics.DataProviders/DataProviders/K8SELogAnalyticsDataProvider.cs
@@ -5,26 +5,23 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Diagnostics.DataProviders
 {
-    public class K8SELogAnalyticsDataProvider : DiagnosticDataProvider, IDiagnosticDataProvider, ILogAnalyticsDataProvider
+    public class K8SELogAnalyticsDataProvider : LogAnalyticsDataProvider
     {
         private ILogAnalyticsClient _K8SELogAnalyticsClient;
+
         public K8SELogAnalyticsDataProvider(OperationDataCache cache, K8SELogAnalyticsDataProviderConfiguration configuration, string RequestId) : base(cache, configuration)
         {
             _K8SELogAnalyticsClient = new K8SELogAnalyticsClient(RequestId);
         }
 
-        public Task<DataTable> ExecuteQueryAsync(string query)
+        public override Task<DataTable> ExecuteQueryAsync(string query)
         {
             return _K8SELogAnalyticsClient.ExecuteQueryAsync(query);
-        }
-
-        public DataProviderMetadata GetMetadata()
-        {
-            return Metadata;
         }
     }
 }
