+++ src/Diagnostics.DataProviders/Interfaces/ILogAnalyticsDataProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System.Data;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Diagnostics.DataProviders
