+++ src/Diagnostics.DataProviders/K8SELogAnalyticsLogDecorator.cs
@@ -1,22 +1,15 @@
 ﻿using System.Data;
+using System.Threading;
 using System.Threading.Tasks;
 using Diagnostics.DataProviders.Interfaces;
 using Diagnostics.ModelsAndUtils.Models;
 
 namespace Diagnostics.DataProviders
 {
-    internal class K8SELogAnalyticsLogDecorator : LogDecoratorBase, ILogAnalyticsDataProvider
+    internal class K8SELogAnalyticsLogDecorator : LogAnalyticsLogDecorator
     {
-        public ILogAnalyticsDataProvider DataProvider;
-
-        public K8SELogAnalyticsLogDecorator(DataProviderContext context, ILogAnalyticsDataProvider dataProvider) : base(context, dataProvider.GetMetadata())
+        public K8SELogAnalyticsLogDecorator(DataProviderContext context, ILogAnalyticsDataProvider dataProvider) : base(context, dataProvider)
         {
-            DataProvider = dataProvider;
-        }
-
-        public Task<DataTable> ExecuteQueryAsync(string query)
-        {
-            return MakeDependencyCall(DataProvider.ExecuteQueryAsync(query));
         }
     }
 }
